CONM-810 Implement ComServerMobile in Connexo - Make Mobile comserver independent of the Oracle Db (test correction in the next commit)

+++ comu/com.energyict.mdc.engine/src/main/java/com/energyict/mdc/engine/impl/core/offline/ComJobExecutionModel.java
@@ -299,7 +299,7 @@ public class ComJobExecutionModel implements CanProvideDescriptionTitle {
         return getSecurityPropertySets();
     }
 
-    public TypedProperties findProtocolDialectPropertiesFor(int comTaskExecutionId) {
+    public TypedProperties findProtocolDialectPropertiesFor(long comTaskExecutionId) {
         return getProtocolDialectPropertiesPerComTaskExecutionIdMap().get(comTaskExecutionId);
     }
